+++ sqlalchemy_utils/functions/orm.py
@@ -724,35 +724,37 @@ def has_changes(obj, attrs=None, exclude=None):
         )
 
 
-def has_any_changes(obj, columns):
+def is_loaded(obj, prop):
     """
-    Simple shortcut function for checking if any of the given attributes of
-    given declarative model object have changes.
-
+    Return whether or not given property of given object has been loaded.
 
     ::
 
-
-        from sqlalchemy_utils import has_any_changes
-
-
-        user = User()
-
-        has_any_changes(user, ('name', ))  # False
-
-        user.name = u'someone'
-
-        has_any_changes(user, ('name', 'age'))  # True
+        class Article(Base):
+            __tablename__ = 'article'
+            id = sa.Column(sa.Integer, primary_key=True)
+            name = sa.Column(sa.String)
+            content = sa.orm.deferred(sa.Column(sa.String))
 
 
-    .. versionadded: 0.26.3
-    .. deprecated:: 0.26.6
-        User :func:`has_changes` instead.
+        article = session.query(Article).get(5)
+
+        # name gets loaded since its not a deferred property
+        assert is_loaded(article, 'name')
+
+        # content has not yet been loaded since its a deferred property
+        assert not is_loaded(article, 'content')
+
+
+    .. versionadded: 0.27.8
 
     :param obj: SQLAlchemy declarative model object
-    :param attrs: Names of the attributes
+    :param prop: Name of the property or InstrumentedAttribute
     """
-    return any(has_changes(obj, column) for column in columns)
+    return not isinstance(
+        getattr(sa.inspect(obj).attrs, prop).loaded_value,
+        sa.util.langhelpers._symbol
+    )
 
 
 def identity(obj_or_class):